backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
django-filter==23.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
phonenumbers==8.13.8
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
